requirements.txt
# requirements.txt — Python worker dependencies
# CPU-only PyTorch: pip install -r requirements.txt
# GPU (CUDA 11.8): pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

uvloop>=0.19.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
prometheus-fastapi-instrumentator>=6.1.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
prometheus-client>=0.18.0
numpy>=1.24.0
scipy>=1.11.0
aiohttp>=3.9.0
